% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirement_pension.R
\name{compute_flat_pension}
\alias{compute_flat_pension}
\title{Compute Flat Pension}
\usage{
compute_flat_pension(dt, params)
}
\arguments{
\item{dt}{data.table. Retiree data}

\item{params}{List with parameter:
\itemize{
\item flat_amount: Fixed pension amount for all retirees
}}
}
\value{
Numeric vector of pension amounts (all identical)
}
\description{
Assigns a uniform flat pension amount to all retirees.
}
\examples{
\dontrun{
params <- list(flat_amount = 15000)
pensions <- compute_flat_pension(retirees_dt, params)
}
}
\keyword{internal}
