% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_date_stamp}
\alias{format_date_stamp}
\title{Format Date as YYYYMMDD String}
\usage{
format_date_stamp(date)
}
\arguments{
\item{date}{Date vector}
}
\value{
Character vector in YYYYMMDD format
}
\description{
Converts dates to compact string format for IDs and timestamps.
}
\keyword{internal}
