% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govhrcast-package.R
\docType{package}
\name{govhrcast-package}
\alias{govhrcast}
\alias{govhrcast-package}
\title{govhrcast: Government Wage Bill Forecasting and Simulation}
\description{
Provides a simulation system for public sector workforce analytics. Allows World Bank country teams to simulate workforce policies, estimate wage bills under baseline or policy scenarios, and model retirement, hiring, promotion, and transfer dynamics using harmonized HRMIS data.
}
\author{
\strong{Maintainer}: Ifeanyi Edochie \email{iedochie@worldbank.org}

Authors:
\itemize{
  \item Galileu Kim \email{galileukim@worldbank.org}
}

}
\keyword{internal}
