% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_shinyutils.R
\name{hz_lever_cols}
\alias{hz_lever_cols}
\title{Identify lever (parameter) columns in a scenario data.table}
\usage{
hz_lever_cols(dt)
}
\arguments{
\item{dt}{A \code{data.table} produced by \code{\link{generate_scenario_matrix}}.}
}
\value{
Character vector of lever column names (may be length 0).
}
\description{
Returns the names of columns that represent policy levers -- i.e. all
columns that are \emph{not} reserved time-series output columns.
}
\keyword{internal}
