% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon_class.R
\name{is.horizon}
\alias{is.horizon}
\title{Test whether an object is a \code{horizon}}
\usage{
is.horizon(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a \code{horizon}
}
