% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirement_core.R
\name{retirement_core}
\alias{retirement_core}
\title{Core Retirement Logic Functions}
\description{
Functions for identifying retirement-eligible personnel and computing
retirement eligibility based on age and/or tenure.
}
\keyword{internal}
