% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_character_string}
\alias{validate_character_string}
\title{Validate Character String}
\usage{
validate_character_string(str, varname, allow_na = FALSE)
}
\arguments{
\item{str}{Object to validate}

\item{varname}{Character. Variable name for error messages}

\item{allow_na}{Logical. Allow NA? (default FALSE)}
}
\value{
Invisible TRUE if valid, stops with error otherwise
}
\description{
Validate Character String
}
\keyword{internal}
