% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_datatable}
\alias{validate_datatable}
\title{Validate Data Table Input}
\usage{
validate_datatable(dt, varname)
}
\arguments{
\item{dt}{Object to validate}

\item{varname}{Character. Variable name for error messages}
}
\value{
Invisible TRUE if valid, stops with error otherwise
}
\description{
Validate Data Table Input
}
\keyword{internal}
