% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_date_format}
\alias{validate_date_format}
\title{Validate Date Format}
\usage{
validate_date_format(date, varname)
}
\arguments{
\item{date}{Object to validate}

\item{varname}{Character. Variable name for error messages}
}
\value{
Invisible TRUE if valid, stops with error otherwise
}
\description{
Validate Date Format
}
\keyword{internal}
