% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon_class.R
\name{validate_horizon}
\alias{validate_horizon}
\title{Validate a \code{horizon} object}
\usage{
validate_horizon(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Invisible \code{x} if valid; stops with informative error otherwise.
}
\description{
Validate a \code{horizon} object
}
\keyword{internal}
