% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_new_ids}
\alias{generate_new_ids}
\title{Generate New Personnel IDs}
\usage{
generate_new_ids(n, ref_date, prefix = "NEW")
}
\arguments{
\item{n}{Integer. Number of IDs to generate}

\item{ref_date}{Date. Reference date for ID stamping}

\item{prefix}{Character. Prefix for IDs (default "NEW")}
}
\value{
Character vector of new IDs in format: PREFIX_YYYYMMDD_N
}
\description{
Creates unique personnel IDs for new hires using a timestamp-based format.
}
\examples{
\dontrun{
generate_new_ids(3, as.Date("2025-01-15"))
# Returns: c("NEW_20250115_1", "NEW_20250115_2", "NEW_20250115_3")
}
}
\keyword{internal}
