% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_active_contracts}
\alias{get_active_contracts}
\title{Get Active Contracts at Reference Date}
\usage{
get_active_contracts(
  contract_dt,
  ref_date,
  start_date_col = "start_date",
  end_date_col = "end_date",
  contract_type_col = "contract_type_code"
)
}
\arguments{
\item{contract_dt}{data.table. Contract data}

\item{ref_date}{Date. Reference date}

\item{start_date_col}{Character. Name of start date column (default: "start_date")}

\item{end_date_col}{Character. Name of end date column (default: "end_date")}

\item{contract_type_col}{Character. Name of contract type column (default: "contract_type_code")}
}
\value{
data.table of active contracts
}
\description{
Filters contracts that are active at a given reference date.
A contract is active if it has started and has not ended before ref_date.
}
\keyword{internal}
