% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_shinyutils.R
\name{hz_app_theme}
\alias{hz_app_theme}
\title{Build the govhrcast app bslib theme}
\usage{
hz_app_theme()
}
\value{
A \code{bs_theme} object.
}
\description{
Returns a \code{bslib::bs_theme()} object with the standard govhrcast
colour palette and typography.
}
\keyword{internal}
