% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirement_core.R
\name{prepare_retiree_data}
\alias{prepare_retiree_data}
\title{Prepare Retiree Data for Pension Calculation}
\usage{
prepare_retiree_data(eligibility_dt, contract_dt, personnel_dt, ref_date)
}
\arguments{
\item{eligibility_dt}{data.table. Output from identify_retirees()}

\item{contract_dt}{data.table. Contract data}

\item{personnel_dt}{data.table. Personnel data}

\item{ref_date}{Date. Reference date}
}
\value{
data.table with retiree information ready for pension calculation
}
\description{
Enriches retiree eligibility data with salary and contract information
needed for pension calculations. Merges age/tenure with contract data.
}
\keyword{internal}
